DTFS2-4518 - split out GEF facility calculations
